// packages/cactus-odap-hermes/src/main/typescript/generated/openapi/typescript-axios/api.ts
export interface SendClientRequestMessage {
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    version: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    loggingProfile: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    accessControlProfile: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    assetControlProfile: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    applicationProfile: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    assetProfile: string;
    /**
     * 
     * @type {object}
     * @memberof SendClientRequestMessage
     */
    payLoadProfile: object;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    sourceGateWayDltSystem: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    recipientGateWayDltSystem: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    recipientGateWayPubkey: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    originatorPubkey: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    beneficiaryPubkey: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    clientIdentityPubkey: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    serverIdentityPubkey: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    clientDltSystem: string;
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessage
     */
    serverDltSystem: string;
    /**
     * 
     * @type {SendClientRequestMessageServerGatewayConfiguration}
     * @memberof SendClientRequestMessage
     */
    serverGatewayConfiguration: SendClientRequestMessageServerGatewayConfiguration;
}
/**
 * 
 * @export
 * @interface SendClientRequestMessageServerGatewayConfiguration
 */
export interface SendClientRequestMessageServerGatewayConfiguration {
    /**
     * 
     * @type {string}
     * @memberof SendClientRequestMessageServerGatewayConfiguration
     */
    apiHost: string;
}